queue and stack moved to common
