+++ strelyaev.roman/S2/calculations.hpp
@@ -3,7 +3,7 @@
 #include <string>
 #include <iosfwd>
 #include "token.hpp"
-#include <otherContainers/queue.hpp>
+#include <queue.hpp>
 namespace strelyaev
 {
   long long calculateOperation(const ExpressionUnit&, const ExpressionUnit&, const ExpressionUnit&);
