+++ strelyaev.roman/S2/conversion.cpp
@@ -3,8 +3,8 @@
 #include <string>
 #include <iomanip>
 #include "calculations.hpp"
-#include <otherContainers/queue.hpp>
-#include <otherContainers/stack.hpp>
+#include <queue.hpp>
+#include <stack.hpp>
 #include "token.hpp"
 
 strelyaev::ExpressionUnit strelyaev::convertStringToUnit(std::string string_token)
