+++ strelyaev.roman/common/stack.hpp
